Support pandas 3 in remaining pandas-consuming libs
- Bridge flags in artifact, fuzzy-checker, gbd-mapping, profiling, and
  risk-distributions test conftests: run pandas >=2.1 suites with
  copy-on-write and future.infer_string enabled (MIC-6773)
- risk-distributions: cast ensemble weights and distribution call data
  to float before transformed values are written into them — pandas 3
  removed silent integer upcasting on setitem, so float-into-int64
  .loc writes now raise TypeError

+++ libs/risk-distributions/src/vivarium/risk_distributions/formatting.py
@@ -216,4 +216,7 @@ def format_call_data(
         call_data = pd.Series(call_data, copy=True)
         parameters = parameters.reindex(call_data.index, method="nearest")
 
-    return call_data, parameters
+    # float: the pdf/ppf pre-processors write transformed (float) values back
+    # into this series, which raises on integer input under pandas 3
+    # (upcasting removed)
+    return call_data.astype(float), parameters
